files - rename disk based file providers to make intent clear
